+++ lib/src/helpers/controller/_exports.dart
@@ -1,4 +1,4 @@
 export 'controller_lifecycle.dart';
 export 'is_loading_mixin.dart';
 export 'loggable_mixin.dart';
-export 'stream_handler_mixin.dart';
+export 'dispose_handler_mixin.dart';